refactor: guideon-common 모듈에 공통 코드 분리 (#29)
