+++ guideon-api/src/main/java/com/guideon/guideonbackend/domain/admin/service/AdminAuthService.java
@@ -11,8 +11,8 @@ import com.guideon.guideonbackend.domain.admin.entity.RefreshToken;
 import com.guideon.guideonbackend.domain.admin.repository.AdminRepository;
 import com.guideon.guideonbackend.domain.admin.repository.AdminSiteRepository;
 import com.guideon.guideonbackend.domain.admin.repository.RefreshTokenRepository;
-import com.guideon.guideonbackend.global.exception.CustomException;
-import com.guideon.guideonbackend.global.exception.ErrorCode;
+import com.guideon.common.exception.CustomException;
+import com.guideon.common.exception.ErrorCode;
 import com.guideon.guideonbackend.global.security.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
