+++ guideon-api/src/main/java/com/guideon/guideonbackend/domain/site/service/SiteService.java
@@ -5,8 +5,8 @@ import com.guideon.guideonbackend.domain.site.dto.SiteResponse;
 import com.guideon.guideonbackend.domain.site.dto.UpdateSiteRequest;
 import com.guideon.guideonbackend.domain.site.entity.Site;
 import com.guideon.guideonbackend.domain.site.repository.SiteRepository;
-import com.guideon.guideonbackend.global.exception.CustomException;
-import com.guideon.guideonbackend.global.exception.ErrorCode;
+import com.guideon.common.exception.CustomException;
+import com.guideon.common.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
