+++ guideon-api/src/main/java/com/guideon/guideonbackend/domain/zone/controller/ZoneController.java
@@ -4,8 +4,8 @@ import com.guideon.guideonbackend.domain.zone.dto.CreateZoneRequest;
 import com.guideon.guideonbackend.domain.zone.dto.DeleteZoneResponse;
 import com.guideon.guideonbackend.domain.zone.dto.ZoneResponse;
 import com.guideon.guideonbackend.domain.zone.service.ZoneService;
-import com.guideon.guideonbackend.global.response.ApiResponse;
-import com.guideon.guideonbackend.global.response.PageResponse;
+import com.guideon.common.response.ApiResponse;
+import com.guideon.common.response.PageResponse;
 import com.guideon.guideonbackend.global.security.CustomAdminDetails;
 import com.guideon.guideonbackend.global.trace.TraceIdUtil;
 import io.swagger.v3.oas.annotations.Operation;
