+++ guideon-api/src/main/java/com/guideon/guideonbackend/domain/zone/service/ZoneService.java
@@ -10,11 +10,11 @@ import com.guideon.guideonbackend.domain.zone.dto.ZoneResponse;
 import com.guideon.guideonbackend.domain.zone.entity.Zone;
 import com.guideon.guideonbackend.domain.zone.entity.ZoneType;
 import com.guideon.guideonbackend.domain.zone.repository.ZoneRepository;
-import com.guideon.guideonbackend.global.exception.CustomException;
-import com.guideon.guideonbackend.global.exception.ErrorCode;
+import com.guideon.common.exception.CustomException;
+import com.guideon.common.exception.ErrorCode;
 import com.guideon.guideonbackend.global.security.CustomAdminDetails;
 import com.guideon.guideonbackend.global.util.GeoJsonUtil;
-import com.guideon.guideonbackend.global.response.PageResponse;
+import com.guideon.common.response.PageResponse;
 import lombok.RequiredArgsConstructor;
 import org.locationtech.jts.geom.Geometry;
 import org.springframework.data.domain.Page;
