--- guideon-api/src/main/java/com/guideon/guideonbackend/global/response/ApiError.java
@@ -1,17 +1,0 @@
-package com.guideon.guideonbackend.global.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-//에러 응답 DTO
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ApiError {
-    private String code;
-    private String message;
-    private Object details;
-}
