+++ guideon-common/src/main/java/com/guideon/common/response/ApiError.java
@@ -1,0 +1,17 @@
+package com.guideon.common.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+//에러 응답 DTO
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ApiError {
+    private String code;
+    private String message;
+    private Object details;
+}
