+++ guideon-common/src/main/java/com/guideon/common/response/PageResponse.java
@@ -1,0 +1,42 @@
+package com.guideon.common.response;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.data.domain.Page;
+
+import java.util.List;
+
+@Getter
+@Builder
+public class PageResponse<T> {
+
+    private List<T> items;
+
+    private PageInfo page;
+
+    public static <T> PageResponse<T> from(Page<T> page) {
+        return PageResponse.<T>builder()
+                .items(page.getContent())
+                .page(PageInfo.builder()
+                        .number(page.getNumber())
+                        .size(page.getSize())
+                        .totalElements(page.getTotalElements())
+                        .totalPages(page.getTotalPages())
+                        .build())
+                .build();
+    }
+
+    @Getter
+    @Builder
+    public static class PageInfo {
+        private int number;
+        private int size;
+
+        @JsonProperty("total_elements")
+        private long totalElements;
+
+        @JsonProperty("total_pages")
+        private int totalPages;
+    }
+}
